+++ modules/personal.nix
@@ -33,40 +33,19 @@
       '';
       overrideStrategy = "asDropin";
     };
-
-    services.resilio.enable = true;
-    services.resilio.listeningPort = 44444;
-    services.resilio.enableWebUI = true;
-    services.resilio.httpListenAddr = "0.0.0.0";
-
-    users.users.${user}.extraGroups = [ "rslsync" ];
   };
 
   homeModule = { config, pkgs, lib, ... }:
     let
       inherit (pkgs.stdenv) hostPlatform;
-      inherit (lib) mkIf optionalAttrs;
+      inherit (lib) optionalAttrs;
     in {
       home.packages = builtins.attrValues (optionalAttrs hostPlatform.isLinux {
         # Currently broken on macOS
         inherit (pkgs) gramps;
-      } // optionalAttrs (!hostPlatform.isLinux || !hostPlatform.isAarch64) {
-        # Runs on everything except `aarch64-linux`
-        discord = if hostPlatform.isLinux then
-          pkgs.discord.override { withOpenASAR = true; }
-        else
-          pkgs.discord;
       } // optionalAttrs (hostPlatform.isLinux && hostPlatform.isx86_64) {
-        inherit (pkgs) joplin-desktop signal-desktop;
-      } // optionalAttrs (hostPlatform.isLinux && hostPlatform.isAarch64) {
-        inherit (pkgs) armcord;
+        inherit (pkgs) joplin-desktop;
       });
-
-      xsession.windowManager.i3.config.startup =
-        mkIf (hostPlatform.isLinux && hostPlatform.isx86_64) [{
-          command = "signal-desktop";
-          always = true;
-        }];
 
       programs.firefox.profiles.personal.isDefault = true;
     };